Code improvements.
