+++ main_test.go
@@ -3,8 +3,9 @@ package main
 import (
 	"bytes"
 	"os"
-	"strings"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 func TestMainApp(t *testing.T) {
@@ -32,20 +33,19 @@ apiVersion: v1
 kind: Namespace
 metadata:
   name: production`,
-			stdout: `# File: pod-nginx-ingress.yaml (58 bytes)
+			stdout: `# File: pod-nginx-ingress.yaml (56 bytes)
 apiVersion: v1
 kind: Pod
 metadata:
   name: nginx-ingress
 
-
 ---
-
-# File: namespace-production.yaml (61 bytes)
+# File: namespace-production.yaml (59 bytes)
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: production`,
+  name: production
+`,
 			stderr: `2 files parsed to stdout.`,
 		},
 		{
@@ -55,15 +55,13 @@ name: bar
 age: baz
 ---
 another: file`,
-			stdout: `# File: foo-.yaml (30 bytes)
+			stdout: `# File: foo-.yaml (28 bytes)
 kind: foo
 name: bar
 age: baz
 
-
 ---
-
-# File: -.yaml (15 bytes)
+# File: -.yaml (13 bytes)
 another: file`,
 			stderr: `2 files parsed to stdout.`,
 		},
@@ -83,17 +81,13 @@ another: file`,
 	for _, c := range cases {
 		t.Run(c.name, func(tt *testing.T) {
 			f, err := os.CreateTemp("/tmp", "kubectl-slice-testing-file-*")
-			if err != nil {
-				tt.Fatalf("unable to create temporary file: %s", err.Error())
-			}
+			require.NoErrorf(tt, err, "unable to create temporary file")
 			defer os.Remove(f.Name())
 
-			if _, err := f.Write([]byte(c.file)); err != nil {
-				tt.Fatalf("unable to write to temporary file: %s", err.Error())
-			}
+			_, err = f.Write([]byte(c.file))
+			require.NoErrorf(tt, err, "unable to write to temporary file")
 
-			var stdout bytes.Buffer
-			var stderr bytes.Buffer
+			var stdout, stderr bytes.Buffer
 
 			baseArgs := []string{"--input-file=" + f.Name(), "--stdout"}
 			args := append(baseArgs, c.flags...)
@@ -102,17 +96,9 @@ another: file`,
 			cmd.SetOut(&stdout)
 			cmd.SetErr(&stderr)
 			cmd.SetArgs(args)
-			if err := cmd.Execute(); err != nil {
-				tt.Fatalf("unable to execute command: %s", err.Error())
-			}
-
-			if strings.TrimSpace(stdout.String()) != c.stdout {
-				tt.Fatalf("stdout mismatch: expected %s, got %s", c.stdout, stdout.String())
-			}
-
-			if strings.TrimSpace(stderr.String()) != c.stderr {
-				tt.Fatalf("stderr mismatch: expected %s, got %s", c.stderr, stderr.String())
-			}
+			require.NoError(tt, cmd.Execute())
+			require.EqualValues(tt, c.stdout, stdout.String())
+			require.EqualValues(tt, c.stderr, stderr.String())
 		})
 	}
 }
